delete innecesari elements
