+++ index.html
@@ -115,7 +115,7 @@
         <p class="p-footer">© GIFOS 2020 All Rights Reserved.</p>
 
     </footer>
-    <script src="./scripts/app1.js" type="module"></script>
+    <script src="./scripts/app.js" type="module"></script>
     
     
 </body>
